Add ability to save new task and edit an existing one
The bug with sortTasks function at display.js that
was fixed a few commits ago, it  wasn't a bug but a feature
that needed some reworking, so it's back and working now. The splice line
was there to ensure the same task wasn't displayed twice if two tasks
had the same date.